Move dts/resources to new resources subpackage
