+++ src/main/scala/amba/axis/Parameters.scala
@@ -7,7 +7,8 @@ import chisel3.util.{isPow2, log2Ceil}
 import org.chipsalliance.cde.config.Parameters
 import org.chipsalliance.diplomacy.nodes.BaseNode
 
-import freechips.rocketchip.diplomacy.{TransferSizes, Resource, IdRange}
+import freechips.rocketchip.diplomacy.{TransferSizes, IdRange}
+import freechips.rocketchip.resources.{Resource}
 import freechips.rocketchip.util.{BundleFieldBase, BundleField}
