+++ src/main/scala/groundtest/Tile.scala
@@ -9,7 +9,7 @@ import org.chipsalliance.cde.config._
 import org.chipsalliance.diplomacy.bundlebridge._
 import org.chipsalliance.diplomacy.lazymodule._
 
-import freechips.rocketchip.diplomacy.{SimpleDevice}
+import freechips.rocketchip.resources.{SimpleDevice}
 import freechips.rocketchip.prci.{ClockCrossingType}
 import freechips.rocketchip.interrupts._
 import freechips.rocketchip.rocket.{BuildHellaCache, DCache, DCacheModule, ICacheParams, NonBlockingDCache, NonBlockingDCacheModule, RocketCoreParams}
