+++ src/main/scala/interrupts/Parameters.scala
@@ -7,7 +7,7 @@ import chisel3.experimental.SourceInfo
 import org.chipsalliance.cde.config._
 import org.chipsalliance.diplomacy.nodes._
 
-import freechips.rocketchip.diplomacy.Resource
+import freechips.rocketchip.resources.Resource
 
 // A potentially empty half-open range; [start, end)
 case class IntRange(start: Int, end: Int)
