+++ src/main/scala/resources/FixedClockResource.scala
@@ -1,6 +1,6 @@
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.diplomacy
+package freechips.rocketchip.resources
 
 class FixedClockResource(val name: String, val freqMHz: Double, val prefix: String = "soc/")
 {
