+++ src/main/scala/resources/SRAM.scala
@@ -1,11 +1,12 @@
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.diplomacy
+package freechips.rocketchip.resources
 
 import chisel3._
 import chisel3.util.log2Ceil
 import org.chipsalliance.cde.config.Parameters
 import freechips.rocketchip.util.{DescribedSRAM, Code}
+import freechips.rocketchip.diplomacy.{AddressSet, LazyModule}
 
 abstract class DiplomaticSRAM(
     val address: AddressSet,
