+++ src/main/scala/subsystem/InterruptBus.scala
@@ -7,7 +7,7 @@ import chisel3._
 import org.chipsalliance.cde.config._
 import org.chipsalliance.diplomacy.lazymodule._
 
-import freechips.rocketchip.diplomacy.{Device, DeviceInterrupts, Description, ResourceBindings}
+import freechips.rocketchip.resources.{Device, DeviceInterrupts, Description, ResourceBindings}
 import freechips.rocketchip.interrupts.{IntInwardNode, IntOutwardNode, IntXbar, IntNameNode, IntSourceNode, IntSourcePortSimple}
 import freechips.rocketchip.prci.{ClockCrossingType, AsynchronousCrossing, RationalCrossing, ClockSinkDomain}
