+++ src/main/scala/tile/Interrupts.scala
@@ -7,12 +7,11 @@ import chisel3.util._
 import org.chipsalliance.cde.config._
 import org.chipsalliance.diplomacy._
 
-import freechips.rocketchip.diplomacy.{Device, DeviceSnippet, Description, ResourceBinding, ResourceInt}
+import freechips.rocketchip.resources.{Device, DeviceSnippet, Description, ResourceBinding, ResourceInt}
 import freechips.rocketchip.interrupts.{IntIdentityNode, IntSinkNode, IntSinkPortSimple, IntSourceNode, IntSourcePortSimple}
 import freechips.rocketchip.util.CanHaveErrors
 
-import freechips.rocketchip.diplomacy.IntToProperty
-import freechips.rocketchip.diplomacy.StringToProperty
+import freechips.rocketchip.resources.{IntToProperty, StringToProperty}
 import freechips.rocketchip.util.BooleanToAugmentedBoolean
 
 class NMI(val w: Int) extends Bundle {
